+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 This project changes will be shown here.
 
+## [unreleased]
+### Added
+- Issue 786 - Create new Furlough Contact Support Level
+
 ## [v3.11.7] - 2020-03-26
 ### Fixed
 - Expenses/Overtime pending approval list